pyproject.toml
[tool.poetry]
name = "workflow"
version = "0.1.0"
description = ""
authors = ["Illia Bahlai <bahlai.illia@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.scripts]
start = "workflow.__main__:run_app"
gen_database = "database_generator.database_factory:gen_database"


[tool.poetry.dependencies]
python = "^3.10"
fastapi = { extras = ["standard"], version = "^0.115.3" }
fastapi-restful = { extras = ["all"], version = "^0.6.0" }
sqlmodel = "^0.0.22"
pydantic-settings = "^2.6.0"
asyncpg = "^0.30.0"
httpx = "^0.27.2"
psycopg2-binary = "^2.9.10"
pydantic-extra-types = "^2.9.0"
pycountry = "^24.6.1"
phonenumbers = "^8.13.48"
tzdata = "^2024.2"
asyncio = "^3.4.3"
greenlet = "^3.1.1"


[tool.poetry.group.dev.dependencies]
mypy = "^1.12.1"
ruff = "^0.7.0"
pre-commit = "^4.0.1"
import-linter = "^2.1"
ipython = "^8.29.0"
sqlacodegen = { git = "git@github.com:agronholm/sqlacodegen.git" }
polyfactory = "^2.17.0"
factory-boy = "^3.3.1"
faker = "^30.8.1"
faker-airtravel = "^0.4"
pytest = "^8.3.3"
types-factory-boy = "^0.4.1"
black = "^24.10.0"
async-factory-boy = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 80
indent-width = 4

lint.ignore = ["E501"]

# Exclude a variety of commonly ignored directories.
# Assume Python 3.10
target-version = "py310"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends", ]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"


[tool.mypy]
plugins = ["pydantic.mypy"]
